feat: Set sample_rate and channel_count via objc interface
Problem: I'm not able to set `channel_count` and `sample_rate` in StreamConfiguration as a consumer of this library.

Solution: Wire up the setter calls to the actual operations to set the aforementioned attributes via objc interface.

Testing: Tested on my Macbook. I'm able to update the aforementioned attributes and see the changes reflected in the channel count and sample rate of the audio samples that I'm receiving in `did_output_sample_buffer` callback.

Reviewer guidance: I am only adding the missing attributes that I need. I'm aware that there are still remaining attributes that are not set via objc interface but decided to only focus on `sample_rate` and `channel_count` attributes in this PR since I have high confidence that they would work from my local testing.

Signed-off-by: Alex Jiao <uohzxela@gmail.com>